aptos-trading-agent/requirements.txt
pandas
numpy
scikit-learn
requests
python-dotenv
joblib
aptos-sdk
